调整结果页展示
