+++ app/data/categories.js
@@ -1,0 +1,44 @@
+const categoriesMap = [
+  {
+    code: 'starters',
+    name: '入门话题',
+    description: '开始了解彼此的基础话题',
+    icon: '💜',
+    status: 1,
+    version: '1.0'
+  },
+  {
+    code: 'relationship',
+    name: '关系',
+    description: '探索你们关系中的各个方面',
+    icon: '💕',
+    status: 1,
+    version: '1.0'
+  },
+  {
+    code: 'close-love',
+    name: '亲密的爱',
+    description: '探讨亲密关系中的期望与感受',
+    icon: '🔞',
+    status: 1,
+    version: '1.0'
+  },
+  {
+    code: 'moral-values',
+    name: '道德与价值观',
+    description: '了解彼此的价值观和道德准则',
+    icon: '🤝',
+    status: 1,
+    version: '1.0'
+  },
+  {
+    code: 'money-finances',
+    name: '金钱与财务',
+    description: '讨论金钱观念和财务规划',
+    icon: '💵',
+    status: 1,
+    version: '1.0'
+  }
+];
+
+module.exports = categoriesMap;
